% Auto-generated: do not edit by hand
\name{''GbifAutosuggest}

\alias{''GbifAutosuggest}

\title{GbifAutosuggest component}

\description{

}

\usage{
''GbifAutosuggest(id=NULL, label=NULL, value=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{label}{Character. A label that will be printed when this component is rendered.}

\item{value}{Logical | numeric | character | named list | unnamed list. The value displayed in the input.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

